Add left-behind CSS
