+++ lxl-web/src/routes/(app)/[[lang=lang]]/[fnurgel=fnurgel]/HoldingsModal.svelte
@@ -288,3 +288,37 @@
 		{/if}
 	</Modal>
 {/if}
+
+<style>
+	.expandable {
+		max-height: var(--max-height);
+		overflow: hidden;
+	}
+
+	.expandable:not(.expanded)::after {
+		height: 3rem;
+		width: 100%;
+		position: absolute;
+		content: '';
+		bottom: 0;
+		left: 0;
+		pointer-events: none;
+		background: linear-gradient(
+			to bottom,
+			--alpha(var(--color-page) / 0%),
+			--alpha(var(--color-page) / 100%)
+		);
+		overflow: hidden;
+	}
+
+	.expanded {
+		max-height: initial;
+	}
+
+	#instance-details {
+		& :global(.contribution-role),
+		& :global(.property-label) {
+			font-size: var(--text-2xs);
+		}
+	}
+</style>
